ASTPrinter: Remove ASTPrinter's dependency on Sema. (#10985)
