+++ include/swift/AST/PrintOptions.h
@@ -24,6 +24,7 @@ namespace swift {
 class GenericEnvironment;
 class CanType;
 class Decl;
+class Pattern;
 class ValueDecl;
 class ExtensionDecl;
 class NominalTypeDecl;
@@ -47,32 +48,6 @@ struct TypeTransformContext {
   NominalTypeDecl *getNominal() const;
 
   bool isPrintingSynthesizedExtension() const;
-};
-
-typedef std::pair<ExtensionDecl*, bool> ExtensionAndIsSynthesized;
-typedef llvm::function_ref<void(ArrayRef<ExtensionAndIsSynthesized>)>
-  ExtensionGroupOperation;
-
-class SynthesizedExtensionAnalyzer {
-  struct Implementation;
-  Implementation &Impl;
-public:
-  SynthesizedExtensionAnalyzer(NominalTypeDecl *Target,
-                               PrintOptions Options,
-                               bool IncludeUnconditional = true);
-  ~SynthesizedExtensionAnalyzer();
-
-  enum class MergeGroupKind : char {
-    All,
-    MergeableWithTypeDef,
-    UnmergeableWithTypeDef,
-  };
-
-  void forEachExtensionMergeGroup(MergeGroupKind Kind,
-                                  ExtensionGroupOperation Fn);
-  bool isInSynthesizedExtension(const ValueDecl *VD);
-  bool shouldPrintRequirement(ExtensionDecl *ED, StringRef Req);
-  bool hasMergeGroup(MergeGroupKind Kind);
 };
 
 class BracketOptions {
@@ -129,6 +104,12 @@ public:
     return kind == K.kind && isType == K.isType;
   }
   bool operator!=(AnyAttrKind K) const { return !(*this == K); }
+};
+
+struct ShouldPrintChecker {
+  virtual bool shouldPrint(const Decl *D, PrintOptions &Options);
+  bool shouldPrint(const Pattern *P, PrintOptions &Options);
+  virtual ~ShouldPrintChecker() = default;
 };
 
 /// Options for printing AST nodes.
@@ -294,6 +275,9 @@ struct PrintOptions {
   /// Whether to print the extensions from conforming protocols.
   bool PrintExtensionFromConformingProtocols = false;
 
+  std::shared_ptr<ShouldPrintChecker> CurrentPrintabilityChecker =
+    std::make_shared<ShouldPrintChecker>();
+
   enum class ArgAndParamPrintingMode {
     ArgumentOnly,
     MatchSource,
@@ -423,6 +407,7 @@ struct PrintOptions {
     return result;
   }
 
+  static PrintOptions printModuleInterface();
   static PrintOptions printTypeInterface(Type T);
 
   void setBaseType(Type T);
@@ -430,6 +415,13 @@ struct PrintOptions {
   void initForSynthesizedExtension(NominalTypeDecl *D);
 
   void clearSynthesizedExtension();
+
+  bool shouldPrint(const Decl* D) {
+    return CurrentPrintabilityChecker->shouldPrint(D, *this);
+  }
+  bool shouldPrint(const Pattern* P) {
+    return CurrentPrintabilityChecker->shouldPrint(P, *this);
+  }
 
   /// Retrieve the print options that are suitable to print the testable interface.
   static PrintOptions printTestableInterface() {
@@ -449,19 +441,7 @@ struct PrintOptions {
 
   /// Retrieve the set of options suitable for interface generation for
   /// documentation purposes.
-  static PrintOptions printDocInterface() {
-    PrintOptions result = PrintOptions::printInterface();
-    result.PrintAccessibility = false;
-    result.SkipUnavailable = false;
-    result.ExcludeAttrList.push_back(DAK_Available);
-    result.ArgAndParamPrinting =
-      PrintOptions::ArgAndParamPrintingMode::BothAlways;
-    result.PrintDocumentationComments = false;
-    result.PrintRegularClangComments = false;
-    result.PrintAccessibility = false;
-    result.PrintFunctionRepresentationAttrs = false;
-    return result;
-  }
+  static PrintOptions printDocInterface();
 
   /// Retrieve the set of options suitable for printing SIL functions.
   static PrintOptions printSIL() {
